remove auth, make analytics home
