+++ app/analytics/page.tsx
@@ -4,11 +4,11 @@ import { getServerSession } from 'next-auth';
 import { redirect } from 'next/navigation';
 
 export default async function AnalyticsPage() {
-  // const session = await getServerSession(authOptions);
+  const session = await getServerSession(authOptions);
 
-  // if (!session) {
-  //   redirect('api/auth/signin');
-  // }
+  if (!session) {
+    redirect('api/auth/signin');
+  }
 
   return <Analytics />;
 }
